feat: implement LED strip data transmission
via Remote Control Peripheral (RMT). This sends out the colors bit by
bit, where one bit is encoded with lows and highs with a certain timing.
The WS2818 spec defines these timings.

The example expects a button connected to GPIO4 and the data line of the
LED strip connected to GPIO2. On button press, 150 RGB values with a
maximum intensity of 20 per color are randomly generated and sent to the
strip via the RMT module.