fix: use static mut array for RMT PulseCodes
this avoids the panic when switching between wifi and render tasks. The
RMT send takes a reference (to now a static array instead of a local)
and that didnt go well when interrupted by the wifi timer, leading to
stack corruption (presumably). Using the static mut array seems to solve
the problem. This adds some unsafe code, but as only one thread ever
works with the array, this should generally be save