feat: introduce beat counting
This includes the API change for LedPattern::beat() to get a BeatCount
object containing the info if the current section of the music bar is
e.g. a quarter or a half. For slicing between full bars and 32th notes.

This can then be used by the patterns to determine, if they need to
react to the beat or not. This enables different patterns to react at a
different speed to the music.

Subsequently, the beat triggering task now runs at a lower wait interval
(8 times more often to trigger at every 32th note).