feat: enum for beat speed input for patterns
this implements a simple enum that can be use as an input to a pattern
to determine how opften it reacts to the given beat.

Strobe: if speed (usize) is 0, the beat is used to trigger changes
ShootingStar: a PatternSpeed variant is passed to the constructor. New
starts are created depending on if the beat matches that speed.